app/requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.10.4
wheel==0.24.0
click==5.1
python-dotenv==0.1.3
oauthlib==1.0.3
requests==2.7.0
requests-oauthlib==0.5.0
